numpad-midi/pyproject.toml
[project]
name = "numpad-midi"
version = "0.1.0"
description = "Numpad MIDI Controller for scale-based music input"
requires-python = ">=3.8"
dependencies = [
    "python-rtmidi>=1.4.9",
    "evdev>=1.6.0",
    "murtaza",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
murtaza = { path = "../python/murtaza" }
